+++ src/code/math/cantor_val2seq.hpp
@@ -13,7 +13,7 @@ inline vec<u32> cantor_val2seq(u64 cv, vec<u64> const &fact) {
   for (u64 i = 0; i < n; cv %= fact[n - i - 1], ++i) s.push_back(u32(cv / fact[n - i - 1]));
   return s;
 }
-inline vec<u32> cantor_val2seq(u64 cv, usz n) { return cantor_val2seq(cv, gen_fact((u32)n, (u64)-1)); }
+inline vec<u32> cantor_val2seq(u64 cv, usz n) { return cantor_val2seq(cv, gen_fact((u32)n, -1_u64)); }
 
 }  // namespace tifa_libs::math
